Commit 8#: geting funky with custom types
in this commit we add some custome types to validat that we have well
formed dat structres for the datasets.  We also go to the additional
trouble to create multiple sites which match different db engines.  this
allows more strict matching but also allows users to check in puppet
which type of dataclass has been passed.