commit #10: lets make things ensureable
When we look at adding this to a class to the profile/role structure
later it could be tempting to make the inclusion optional with something
like

```puppet
if $enable_myfunckyservice {
  include myfunckyservice
}
```

This allows on to only enable myfunckyservice on some hosts based on the
boolean value of `$enable_myfunckyservice`.  This works fine for slowly
introducing a service however it dosn;t actully allow is to disable the
service after the fact.  Setting `$enable_myfunckyservice = false` will
simply mean that the class is not included in the manifest and that any
resources it manages are no longer managed.  however
 * any started services will still be started
 * installed packages will still be installed
 * and any files created will still exists

Makeing a class ensureable $ensure fixes this issue.  We make a class
ensureable by adding and implementing the $ensure parameter.  In most
cases we can either pass the ensure parameter directly to all/most
managed resources or use the `stdlib::ensure` function for more
flexibility